requirements.txt
sentence-transformers==2.5.1
faiss-cpu==1.8.0
python-docx==1.1.0
numpy==1.26.4
pathlib==1.0.1
tqdm==4.66.2
